All recomendations are taken
