+++ celery_worker.py
@@ -1,17 +1,11 @@
 from celery import Celery
-from dotenv import load_dotenv
+from config import redis_host, redis_port, redis_db, api_key
 import json
-import os
 from psql import save_currency_rate
 import requests
 import redis
 
 
-# Подключение к Redis
-load_dotenv()
-redis_host = os.getenv('REDIS_HOST')
-redis_port = os.getenv('REDIS_PORT')
-redis_db = os.getenv('REDIS_DB')
 redis_instance = redis.StrictRedis(host=redis_host, port=int(redis_port), db=int(redis_db))
 # Настройка Celery для использования Redis
 celery_app = Celery("currency_worker", broker=f"redis://{redis_host}:{redis_port}/{redis_db}")
@@ -21,8 +15,7 @@ key_currencies = ('RUB', 'EUR', 'BTC', 'CNY', 'USD')
 
 @celery_app.task
 def fetch_currency_data():
-    key = os.getenv('KEY')
-    headers = {'apikey': key}
+    headers = {'apikey': api_key}
     res = requests.get('https://api.currencyapi.com/v3/latest', headers=headers)
     data = res.json()
     exchange_rates = {}
@@ -31,13 +24,13 @@ def fetch_currency_data():
         save_currency_rate(currency, rate)  # Сохранение в PostgreSQL
         exchange_rates[currency] = rate
     # Сохранение курсов валют в Redis
-    redis_instance.set("currency_data", json.dumps(exchange_rates), ex=6)  # 10 минут в секундах
+    redis_instance.set("currency_data", json.dumps(exchange_rates), ex=15)  # 10 минут в секундах
 
 
 # Настройка расписания для Celery
 celery_app.conf.beat_schedule = {
     "fetch-currency-data-every-15-minutes": {
         "task": "celery_worker.fetch_currency_data",
-        "schedule": 9.0,  # 15 минут в секундах
+        "schedule": 30.0,  # 15 минут в секундах
     }
 }
