+++ models/models.py
@@ -1,12 +1,12 @@
-from sqlalchemy import MetaData, Integer, String, Float, TIMESTAMP, Table, Column
-import datetime
+from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy import Column, Integer, String, Float, TIMESTAMP
+from datetime import datetime
 
-metadata = MetaData()
-currency_rates = Table(
-    "currency_rates",
-    metadata,
-    Column("id", Integer, primary_key=True),
-    Column("currency", String, nullable=False),
-    Column("rate", Float),
-    Column("datetime", TIMESTAMP, default=datetime.datetime.utcnow)
-)
+Base = declarative_base()
+
+class CurrencyRates(Base):
+    __tablename__ = 'currency_rates'
+    id = Column(Integer, primary_key=True)
+    currency = Column(String, nullable=False)
+    rate = Column(Float)
+    datetime = Column(TIMESTAMP, default=datetime.utcnow)
